refactor(auth):  clean up and modif archi project
