+++ backend/services/auth/internal/handlers/auth_account.go
@@ -1,7 +1,7 @@
 package handlers
 
 import (
-	"auth/backend/internal/models"
+	"auth/backend/services/auth/internal/models"
 	"encoding/hex"
 	"log"
