+++ backend/services/auth/internal/handlers/auth_crypto.go
