+++ backend/services/auth/internal/models/user.go
