+++ backend/services/auth/internal/repository/user_repository.go
